refactor plugin.test.ts into smaller files
